moves reused components to generic components folder
