+++ src/components/CellphoneMockup/CellphoneMockup.css.ts
@@ -1,5 +1,5 @@
 import { style } from "@vanilla-extract/css";
-import { vars } from "../../../../styles/theme.css";
+import { vars } from "../../styles/theme.css";
 
 export const container = style({
     position: "relative",
