+++ src/components/PreviewCard/PreviewCard.css.ts
@@ -1,6 +1,6 @@
 import { style } from "@vanilla-extract/css";
 import { recipe } from "@vanilla-extract/recipes";
-import { vars } from "../../../../styles/theme.css";
+import { vars } from "../../styles/theme.css";
 
 export const previewCard = style({
   width: "45%",
