+++ src/components/PreviewCard/index.ts
