+++ src/pages/LinksTab/LinksTab.tsx
@@ -1,9 +1,9 @@
 import { useEffect } from "react";
 import { useFormContext } from "react-hook-form";
+import { PreviewCard } from "../../components/PreviewCard";
 import { useGetLinks } from "../../services/LinksService/useGetLinks";
 import { linksTabContainer } from "./LinksTab.css";
 import { CustomizeLinksCard } from "./components/CustomizeLinksCard";
-import { PreviewCard } from "./components/PreviewCard";
 
 
 export const LinksTab = () => {
