--- src/pages/ProfileDetailsTab/ProfileDetailsTab.css
@@ -1,12 +1,0 @@
-import {
-    style
-}
-
-from '@vanilla-extract/css';
-
-export const footerContainer=style({
-    display: 'flex',
-    justifyContent: 'flex-end',
-    width: '432px',
-    marginTop: '24px'
-});
